requirements.txt
alabaster==0.7.12
attrs==19.1.0
Babel==2.7.0
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
dash==1.0.2
dash-core-components==1.0.0
dash-daq==0.1.7
dash-html-components==1.0.0
dash-renderer==1.0.0
dash-table==4.0.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.3
django-analytical==2.5.0
django-crispy-forms==1.7.2
django-import-export==1.2.0
django-reversion==3.0.4
django-reversion-compare==0.8.6
docutils==0.15.2
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Compress==1.4.0
idna==2.8
imagesize==1.1.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
lxml==4.4.0
MarkupSafe==1.1.1
numpy==1.17.0
odfpy==1.4.0
openpyxl==2.6.3
packaging==19.1
pandas==0.25.0
pandas-datareader==0.7.4
Pillow==6.1.0
plotly==4.0.0
psycopg2==2.8.3
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
retrying==1.3.3
six==1.12.0
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
